backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# PDF extraction
pdfplumber==0.11.0
PyMuPDF==1.24.3

# Data processing
pandas==2.2.2
openpyxl==3.1.2   # Excel support for pandas
xlrd==2.0.1       # Old .xls support

# OCR (optional — comment out if Tesseract not installed)
pytesseract==0.3.10
Pillow==10.3.0
